+++ sdk/search/azure-search-documents/tests/test_service_live.py
@@ -17,7 +17,7 @@ from azure.core import MatchConditions
 from azure.core.credentials import AzureKeyCredential
 from azure.core.exceptions import HttpResponseError
 from azure.search.documents.indexes.models import(
-    AnalyzeTextRequest,
+    AnalyzeTextOptions,
     AnalyzeResult,
     CorsOptions,
     EntityRecognitionSkill,
@@ -242,7 +242,7 @@ class SearchIndexesClientTest(AzureMgmtTestCase):
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     def test_analyze_text(self, api_key, endpoint, index_name, **kwargs):
         client = SearchIndexClient(endpoint, AzureKeyCredential(api_key))
-        analyze_request = AnalyzeTextRequest(text="One's <two/>", analyzer="standard.lucene")
+        analyze_request = AnalyzeTextOptions(text="One's <two/>", analyzer_name="standard.lucene")
         result = client.analyze_text(index_name, analyze_request)
         assert len(result.tokens) == 2